+++ cst/plugins/eu.esdihumboldt.cst.functions.core/src/eu/esdihumboldt/cst/functions/core/Merge.java
@@ -16,10 +16,18 @@
 
 package eu.esdihumboldt.cst.functions.core;
 
-import eu.esdihumboldt.cst.functions.core.merge.PropertiesMergeHandler;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+
+import eu.esdihumboldt.cst.functions.core.merge.IndexMergeHandler;
+import eu.esdihumboldt.hale.common.align.model.Cell;
 import eu.esdihumboldt.hale.common.align.model.functions.MergeFunction;
+import eu.esdihumboldt.hale.common.align.model.functions.merge.MergeUtil;
+import eu.esdihumboldt.hale.common.align.model.impl.PropertyEntityDefinition;
 import eu.esdihumboldt.hale.common.align.transformation.engine.TransformationEngine;
 import eu.esdihumboldt.hale.common.align.transformation.function.InstanceHandler;
+import eu.esdihumboldt.hale.common.instance.index.InstanceIndexContribution;
 
 /**
  * Type transformation that merges multiple instances of the same source type
@@ -27,13 +35,31 @@ import eu.esdihumboldt.hale.common.align.transformation.function.InstanceHandler
  * 
  * @author Simon Templer
  */
-public class Merge extends Retype implements MergeFunction {
+public class Merge extends Retype implements MergeFunction, InstanceIndexContribution {
 
 	/**
 	 * @see eu.esdihumboldt.hale.common.align.transformation.function.impl.AbstractTypeTransformation#getInstanceHandler()
 	 */
 	@Override
 	public InstanceHandler<? super TransformationEngine> getInstanceHandler() {
-		return new PropertiesMergeHandler();
+		return new IndexMergeHandler();
+	}
+
+	/**
+	 * @see eu.esdihumboldt.hale.common.instance.index.InstanceIndexContribution#getIndexContribution(eu.esdihumboldt.hale.common.align.model.Cell)
+	 */
+	@Override
+	public Collection<List<PropertyEntityDefinition>> getIndexContribution(Cell cell) {
+		return getMergeContribution(cell);
+	}
+
+	/**
+	 * Calculates the instance index contribution for the given cell
+	 * 
+	 * @param cell Cell
+	 * @return The properties to index
+	 */
+	public static Collection<List<PropertyEntityDefinition>> getMergeContribution(Cell cell) {
+		return Collections.singletonList(MergeUtil.getKeyPropertyDefinitions(cell));
 	}
 }